Add unit tests for the embedding operation
This patch adds C++ unit tests for embedding operation/options.
