+++ test/embedding_test.cc
@@ -1,5 +1,5 @@
 #define BOOST_TEST_DYN_LINK
-#define BOOST_TEST_MODULE quadpool
+#define BOOST_TEST_MODULE embedding
 
 #include <boost/test/included/unit_test.hpp>
 #include <torch/torch.h>
@@ -144,26 +144,12 @@ BOOST_AUTO_TEST_CASE(embedding2d_eval_large)
     auto options = torch::TensorOptions().dtype(torch::kFloat64).device(torch::kCPU);
 
     auto weight = torch::rand({1024, 1024, 3}, options);
-    auto input = torch::rand({100, 2}, options) * 10.0;
+    auto input = torch::rand({100, 2}, options) * 12.0;
 
     auto output
         = embedding2d(input, weight, /*padding=*/{3, 2}, /*exterior=*/{-10.0, 10.0, 20.0, 20.0});
 
     BOOST_CHECK_EQUAL(output.sizes(), c10::IntArrayRef({100, 7, 5, 3}));
-}
-
-
-BOOST_AUTO_TEST_CASE(embedding2d_edge_input)
-{
-    auto options = torch::TensorOptions().dtype(torch::kFloat64).device(torch::kCPU);
-
-    auto weight = torch::rand({1024, 1024, 3}, options);
-    auto input = torch::tensor({{9.98656, 2.8601}}, options);
-
-    auto output
-        = embedding2d(input, weight, /*padding=*/{3, 2}, /*exterior=*/{-10.0, 10.0, 20.0, 20.0});
-
-    BOOST_CHECK_EQUAL(output.sizes(), c10::IntArrayRef({1, 7, 5, 3}));
 }
